Versão 0.9.2 (Calculo porcentagem corrigido "Papelaria")

+++ src/main/java/com/api/papeis/cores/API_Papeis/Cores/Controller/JavaFX/Estoque/Papelaria/ControllerAddPapelaria.java
@@ -176,7 +176,7 @@ public class ControllerAddPapelaria implements Initializable {
             Double precoPacote = Double.parseDouble(textFieldPrecoPacote.getText());
             Integer qntCadaPacote = Integer.parseInt(textFieldQntPacote.getText());
             Double lucro = Double.parseDouble(textFieldLucro.getText());
-            Integer margem = Integer.parseInt(textFieldMargem.getText());
+            Double margem = Double.parseDouble(textFieldMargem.getText());
 
 
             // CHATGPT COMO EU CONSIGO SOMAR O valorTotal e passa como parametro na classe papelaria?
